+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @stackframe/stack-dashboard
 
+## 2.5.2
+
+### Patch Changes
+
+- Team profile pictures
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.2
+  - @stackframe/stack@2.5.2
+  - @stackframe/stack-emails@2.5.2
+
+## 2.5.1
+
+### Patch Changes
+
+- New backend endpoints
+- Updated dependencies
+  - @stackframe/stack-emails@2.5.1
+  - @stackframe/stack-shared@2.5.1
+  - @stackframe/stack@2.5.1
+
 ## 2.5.0
 
 ### Minor Changes