+++ examples/cjs-test/CHANGELOG.md
@@ -1,5 +1,20 @@
 # cjs-test
 
+## 2.5.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.5.2
+
+## 2.5.1
+
+### Patch Changes
+
+- New backend endpoints
+- Updated dependencies
+  - @stackframe/stack@2.5.1
+
 ## 2.5.0
 
 ### Patch Changes