+++ examples/docs-examples/CHANGELOG.md
@@ -1,5 +1,23 @@
 # demo-app
 
+## 2.5.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.2
+  - @stackframe/stack@2.5.2
+  - @stackframe/stack-ui@2.5.2
+
+## 2.5.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.1
+  - @stackframe/stack@2.5.1
+  - @stackframe/stack-ui@2.5.1
+
 ## 2.5.0
 
 ### Minor Changes